+++ src/CMakeLists.txt
@@ -57,6 +57,7 @@ lua_source(lua_sources lua/csv.lua csv_lua)
 lua_source(lua_sources lua/strict.lua strict_lua)
 lua_source(lua_sources lua/clock.lua clock_lua)
 lua_source(lua_sources lua/title.lua title_lua)
+lua_source(lua_sources lua/utils.lua utils_lua)
 lua_source(lua_sources lua/argparse.lua argparse_lua)
 lua_source(lua_sources lua/env.lua env_lua)
 lua_source(lua_sources lua/pwd.lua pwd_lua)
