+++ src/box/lua/schema.lua
@@ -9,6 +9,12 @@ local fiber = require('fiber')
 local session = box.session
 local internal = box.internal
 local utf8 = require('utf8')
+local utils = require('internal.utils')
+
+local check_param = utils.check_param
+local check_param_table = utils.check_param_table
+local update_param_table = utils.update_param_table
+
 local function setmap(table)
     return setmetatable(table, { __serialize = 'map' })
 end
@@ -232,119 +238,6 @@ local function revoke_object_privs(object_type, object_id)
         local uid = tuple.grantee
         _priv:delete{uid, object_type, object_id}
     end
-end
-
--- Same as type(), but returns 'number' if 'param' is
--- of type 'cdata' and represents a 64-bit integer.
-local function param_type(param)
-    local t = type(param)
-    if t == 'cdata' and tonumber64(param) ~= nil then
-        t = 'number'
-    end
-    return t
-end
-
---[[
- @brief Common function to check table with parameters (like options)
- @param table - table with parameters
- @param template - table with expected types of expected parameters
-  type could be comma separated string with lua types (number, string etc),
-  or 'any' if any type is allowed. Instead of this string, function, which will
-  be used to check if the parameter is correct, can be used too. It should
-  accept option as an argument and return either true or false + expected_type.
- The function checks following:
- 1)that parameters table is a table (or nil)
- 2)all keys in parameters are present in template
- 3)type of every parameter fits (one of) types described in template
- The functions calls box.error(box.error.ILLEGAL_PARAMS, ..) on error
- @example
- check_param_table(options, { user = 'string',
-                              port = 'string, number',
-                              data = 'any',
-                              addr = function(opt)
-                                if not ffi.istype(addr_t, buf) then
-                                    return false, "struct addr"
-                                end
-                                return true
-                              end} )
---]]
-local function check_param_table(table, template)
-    if table == nil then
-        return
-    end
-    if type(table) ~= 'table' then
-        box.error(box.error.ILLEGAL_PARAMS,
-                  "options should be a table")
-    end
-    for k,v in pairs(table) do
-        if template[k] == nil then
-            box.error(box.error.ILLEGAL_PARAMS,
-                      "unexpected option '" .. k .. "'")
-        elseif type(template[k]) == 'function' then
-            local res, expected_type = template[k](v)
-            if not res then
-                box.error(box.error.ILLEGAL_PARAMS,
-                          "options parameter '" .. k ..
-                          "' should be of type " .. expected_type)
-            end
-        elseif template[k] == 'any' then -- luacheck: ignore
-            -- any type is ok
-        elseif (string.find(template[k], ',') == nil) then
-            -- one type
-            if param_type(v) ~= template[k] then
-                box.error(box.error.ILLEGAL_PARAMS,
-                          "options parameter '" .. k ..
-                          "' should be of type " .. template[k])
-            end
-        else
-            local good_types = string.gsub(template[k], ' ', '')
-            local haystack = ',' .. good_types .. ','
-            local needle = ',' .. param_type(v) .. ','
-            if (string.find(haystack, needle) == nil) then
-                box.error(box.error.ILLEGAL_PARAMS,
-                          "options parameter '" .. k ..
-                          "' should be one of types: " .. template[k])
-            end
-        end
-    end
-end
-box.internal.check_param_table = check_param_table
-
---[[
- @brief Common function to check type parameter (of function)
- Calls box.error(box.error.ILLEGAL_PARAMS, ) on error
- @example: check_param(user, 'user', 'string')
---]]
-local function check_param(param, name, should_be_type)
-    if param_type(param) ~= should_be_type then
-        box.error(box.error.ILLEGAL_PARAMS,
-                  name .. " should be a " .. should_be_type)
-    end
-end
-box.internal.check_param = check_param
-
---[[
- Adds to a table key-value pairs from defaults table
-  that is not present in original table.
- Returns updated table.
- If nil is passed instead of table, it's treated as empty table {}
- For example update_param_table({ type = 'hash', temporary = true },
-                                { type = 'tree', unique = true })
-  will return table { type = 'hash', temporary = true, unique = true }
---]]
-local function update_param_table(table, defaults)
-    local new_table = {}
-    if defaults ~= nil then
-        for k,v in pairs(defaults) do
-            new_table[k] = v
-        end
-    end
-    if table ~= nil then
-        for k,v in pairs(table) do
-            new_table[k] = v
-        end
-    end
-    return new_table
 end
 
 local function feedback_save_event(event)
