+++ src/lua/init.c
@@ -135,6 +135,7 @@ extern char minifio_lua[],
 	tap_lua[],
 	fio_lua[],
 	error_lua[],
+	utils_lua[],
 	argparse_lua[],
 	iconv_lua[],
 	/* jit.* library */
@@ -302,6 +303,7 @@ static const char *lua_modules[] = {
 	"tap", tap_lua,
 	"help.en_US", help_en_US_lua,
 	"help", help_lua,
+	"internal.utils", utils_lua,
 	"internal.argparse", argparse_lua,
 	"internal.trigger", trigger_lua,
 	"pwd", pwd_lua,
